feat: start of implementing ECS in Game Engine

+++ GameEngine/include/Entity/IEntity.hpp
@@ -13,7 +13,7 @@ class IEntity {
     public:
         virtual bool isPlayer() = 0;
         virtual sf::Sprite getSprite() = 0;
-        virtual void setSpritePosition(int x, int y) = 0;
+        virtual void setSpritePosition(float x, float y) = 0;
         virtual int getSourceRect() = 0;
         virtual void setSourceRect(int rect) = 0;
         virtual void updateTextureRect() = 0;
